Alot has happened to this server code
I did not finish lab 2 so I put in the posted code. Then did the requirements for assignment 2 which you can specifically see in the connection loop when I receive positionData from the client . In the Game loop I sent the position data to all players. Also updated the packets per second to 30!